# tests: stop testing 'express-graphql' instrumentation