--- a/.ci/tav.json
+++ b/.ci/tav.json
@@ -12,7 +12,6 @@
     "cassandra-driver",
     "elasticsearch",
     "express",
-    "express-graphql",
     "express-queue",
     "fastify",
     "finalhandler",