--- a/.npmrc
+++ b/.npmrc
@@ -1,4 +1,4 @@
-# Workaround unresolvable peerDependencies between express-graphql, graphql,
+# Workaround unresolvable peerDependencies between graphql, @apollo-server,
 # and apollo-server-express. npm v7 (included with node v15) makes these
 # peerDependencies issues an install error. Until the community catches up
 # and resolves peerDependencies issues or apm-agent-nodejs.git's tests are