--- a/docs/configuration.asciidoc
+++ b/docs/configuration.asciidoc
@@ -1008,7 +1008,7 @@ Example using options object:
 [source,js]
 ----
 require('elastic-apm-node').start({
-  disableInstrumentations: ['graphql', 'express-graphql']
+  disableInstrumentations: ['graphql', 'redis']
 })
 ----
 
@@ -1016,7 +1016,7 @@ Example using environment variable:
 
 [source,bash]
 ----
-ELASTIC_APM_DISABLE_INSTRUMENTATIONS=graphql,express-graphql
+ELASTIC_APM_DISABLE_INSTRUMENTATIONS=graphql,redis
 ----
 
 For an always up-to-date list of modules for which instrumentation can be disabled,