--- a/lib/instrumentation/modules/graphql.js
+++ b/lib/instrumentation/modules/graphql.js
@@ -140,8 +140,8 @@ module.exports = function (graphql, agent, { version, enabled }) {
         span.name = 'GraphQL: ' + (operationName ? operationName + ' ' : '') + queries.join(', ')
       }
 
-      // `_graphqlRoute` is a boolean set in instrumentations of other modules
-      // that specify 'graphql' in peerDependencies (e.g. 'express-graphql') to
+      // `_graphqlRoute` is a boolean, set in instrumentations of other modules
+      // that specify 'graphql' in peerDependencies (e.g. '@apollo/server') to
       // indicate that this transaction is for a GraphQL request.
       const trans = span.transaction
       if (trans._graphqlRoute) {