--- a/package.json
+++ b/package.json
@@ -157,7 +157,6 @@
     "eslint-plugin-promise": "^4.3.1",
     "eslint-plugin-standard": "^4.1.0",
     "express": "^4.17.1",
-    "express-graphql": "^0.12.0",
     "express-queue": "^0.0.13",
     "fastify": "^4.16.3",
     "finalhandler": "^1.1.2",