… examples/trace-express-graphql.js
… test/instrumentation/modules/express-graphql.test.js