@trentm 2023-05-01
Some timing info running the graphql TAV tests locally on my dev machine:

```
# before
TAV=graphql ./node_modules/.bin/tav --quiet  394.87s user 81.70s system 109% cpu 7:15.02 total

# after
TAV=graphql ./node_modules/.bin/tav --quiet  213.31s user 47.51s system 90% cpu 4:47.81 total
```

So, yes this is a large savings.
(Seperately we could also reduce the number of graphql versions tested by
just testing the latest patch in each minor release.)